Introduce stringrange.h/cpp and visualstudio.h/cpp
